# Minor optimization